IGNITE-17905 Move Rest, Compute and Client configuration schemas
